refactor: adjust code for new db schemas

+++ src/error.rs
@@ -124,7 +124,7 @@ impl Error {
 	/// When PATCHing maps, the user shouldn't be allowed to update courses that do not belong
 	/// to the map.
 	#[track_caller]
-	pub(crate) fn course_does_not_belong_to_map(course_id: u32, map_id: u16) -> Self {
+	pub(crate) fn course_does_not_belong_to_map(course_id: u16, map_id: u16) -> Self {
 		Self::new(StatusCode::CONFLICT).with_message(format_args!(
 			"course with ID `{course_id}` does not belong to map `{map_id}`"
 		))
@@ -133,7 +133,7 @@ impl Error {
 	/// When PATCHing maps, the user shouldn't be allowed to update filters that do not belong
 	/// to courses on the map.
 	#[track_caller]
-	pub(crate) fn filter_does_not_belong_to_course(filter_id: u32, course_id: u32) -> Self {
+	pub(crate) fn filter_does_not_belong_to_course(filter_id: u16, course_id: u16) -> Self {
 		Self::new(StatusCode::CONFLICT).with_message(format_args!(
 			"filter with ID `{filter_id}` does not belong to course `{course_id}`"
 		))
