+++ src/players/models.rs
@@ -105,20 +105,20 @@ pub struct Session {
 
 	/// More data grouped by course & mode.
 	#[serde(deserialize_with = "Session::deserialize_course_sessions")]
-	pub course_sessions: BTreeMap<u32, CourseSession>,
+	pub course_sessions: BTreeMap<u16, CourseSession>,
 }
 
 impl Session {
 	/// Deserializes and validates submitted course sessions.
 	fn deserialize_course_sessions<'de, D>(
 		deserializer: D,
-	) -> Result<BTreeMap<u32, CourseSession>, D::Error>
+	) -> Result<BTreeMap<u16, CourseSession>, D::Error>
 	where
 		D: Deserializer<'de>,
 	{
 		use serde::de;
 
-		let course_sessions = BTreeMap::<u32, CourseSession>::deserialize(deserializer)?;
+		let course_sessions = BTreeMap::<u16, CourseSession>::deserialize(deserializer)?;
 
 		if let Some(course_id) = course_sessions
 			.iter()
